+++ src/components/dashboard/sidebar.tsx
@@ -7,17 +7,17 @@ import { PanelLeft } from "lucide-react";
 
 import { useIsMobile } from "@/hooks/use-mobile";
 import { cn } from "@/lib/utils";
-import { Button } from "@/components/ui/button_sidebar";
-import { Input } from "@/components/ui/input_sidebar";
-import { Separator } from "@/components/ui/separator";
-import { Sheet, SheetContent } from "@/components/ui/sheet";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Button } from "@/components/dashboard/button_sidebar";
+import { Input } from "@/components/dashboard/input_sidebar";
+import { Separator } from "@/components/dashboard/separator";
+import { Sheet, SheetContent } from "@/components/dashboard/sheet";
+import { Skeleton } from "@/components/dashboard/skeleton";
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@/components/ui/tooltip";
+} from "@/components/dashboard/tooltip";
 
 const SIDEBAR_COOKIE_NAME = "sidebar:state";
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7;
